+++ crates/punktf-lib/src/template/resolve.rs
@@ -107,6 +107,13 @@ pub struct Resolver<'a, PV, DV> {
 	/// Session where all errors/diagnostic which occur during the resolving
 	/// process are recorded to.
 	session: Session,
+
+	/// Flag that when it is set prevents a leading new line of a text block to
+	/// be emitted.
+	///
+	/// This is implemented to avoid extra empty lines which could be created
+	/// by either a `comment`, `escaped`, `print` or a not taken `if` block.
+	should_skip_next_newline: bool,
 }
 
 impl<'a, PV, DV> Resolver<'a, PV, DV>
@@ -126,6 +133,7 @@ where
 			profile_vars,
 			dotfile_vars,
 			session: Session::new(),
+			should_skip_next_newline: false,
 		}
 	}
 
@@ -173,20 +181,67 @@ where
 	) -> Result<(), DiagnosticBuilder> {
 		let Block { span, kind } = block;
 
+		// NOTE: `self.should_skip_next_newline` must be reset on every match
+		// branch which is not a text block to avoid remembering it even though
+		// there already was a new non text block between the setting block and
+		// an eventual text block.
 		match kind {
 			BlockKind::Text => {
-				output.push_str(&self.template.source[span]);
+				let mut content = &self.template.source[span];
+
+				// If last block was a block which inserted no content and
+				// started at the beginning of a new line THEN strip a leading
+				// `\n` from the content.
+				// (related #64)
+				if self.should_skip_next_newline
+					&& matches!(content.as_bytes(), &[b'\n', ..] | &[b'\r', b'\n', ..])
+				{
+					// In the if expression above we already checked that there
+					// is an new line character. So if the find fails we messed
+					// something up or the std lib has an error.
+					let lf_idx = content
+						.find('\n')
+						.expect("Failed to find new line character");
+
+					content = &content[lf_idx + 1..];
+
+					self.should_skip_next_newline = false;
+				}
+
+				output.push_str(content);
 			}
 			BlockKind::Comment => {
+				// Should skip new line if started at the beginning of a line.
+				// As a `print` block has no final `content` is the above the
+				// only condition.
+				self.should_skip_next_newline =
+					self.template.source.get_pos_location(span.low).column() == 0;
+
 				// NOP
 			}
 			BlockKind::Escaped(inner) => {
-				output.push_str(&self.template.source[inner]);
+				let content = &self.template.source[inner];
+
+				// Should skip new line if started at the beginning of a line.
+				// As a `print` block has no final `content` is the above the
+				// only condition.
+				self.should_skip_next_newline = content.is_empty()
+					&& self.template.source.get_pos_location(span.low).column() == 0;
+
+				output.push_str(content);
 			}
 			BlockKind::Var(var) => {
+				self.should_skip_next_newline = false;
+
 				output.push_str(&self.resolve_var(var)?);
 			}
 			BlockKind::Print(inner) => {
+				// Should skip new line if started at the beginning of a line.
+				// As a `print` block has no final `content` is the above the
+				// only condition.
+				self.should_skip_next_newline =
+					self.template.source.get_pos_location(span.low).column() == 0;
+
 				log::info!("Print: {}", &self.template.source[inner]);
 			}
 			BlockKind::If(If {
@@ -263,6 +318,11 @@ where
 						if_output_prepared = &if_output_prepared[..idx];
 					}
 				}
+
+				// Should skip new line if started at the beginning of a line
+				// and no new content was added.
+				self.should_skip_next_newline = if_output_prepared.is_empty()
+					&& self.template.source.get_pos_location(span.low).column() == 0;
 
 				output.push_str(if_output_prepared);
 			}
@@ -420,20 +480,149 @@ there
 {{@fi}}
 !"#,
 			"Hello\n\n\tthere\n!"
-		)
+		),
+		(
+			r#"{{@if !{{OS}}}}
+Hello World
+{{@fi}}
+
+Hello
+"#,
+			r#"Hello World
+
+Hello
+"#
+		),
+		// BUG #64: https://github.com/Shemnei/punktf/issues/64
+		(
+			r#"{{@if {{OS}}}} Hello World {{@fi}}
+Hello
+"#,
+			r#"Hello
+"#
+		),
+		(
+			r#"{{@if {{OS}}}}
+	Hello World
+{{@fi}}
+Hello
+"#,
+			r#"Hello
+"#
+		),
+		(
+			r#"{{@if {{OS}}}}
+	Hello World
+{{@fi}}
+
+Hello
+"#,
+			r#"
+Hello
+"#
+		),
+		(
+			r#"Hello
+
+{{@if {{OS}}}}
+	Hello World
+{{@fi}}
+
+World"#,
+			r#"Hello
+
+
+World"#
+		),
+		(
+			r#"{{@print Hello World}}
+
+Hello
+"#,
+			r#"
+Hello
+"#
+		),
+		(
+			r#"{{@if {{OS}}}}
+	Hello World
+{{@fi}}
+
+{{DEMO_VAR}}
+"#,
+			r#"
+DEMO
+"#
+		),
+		(
+			r#"Hello
+{{@if {{OS}}}}Hello World{{@fi}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello
+{{@print Hello World}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello
+{{{}}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello
+{{!-- Comment --}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello{{@if {{OS}}}}Hello World{{@fi}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello{{@print Hello World}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello{{{}}}
+World"#,
+			r#"Hello
+World"#
+		),
+		(
+			r#"Hello{{!-- Comment --}}
+World"#,
+			r#"Hello
+World"#
+		),
 	];
 
 	#[test]
 	fn if_fmt() -> Result<()> {
 		crate::tests::setup_test_env();
 
+		let vars = Variables::from_items([("DEMO_VAR", "DEMO")]);
+
 		for (content, should) in IF_FMT_TEST_CASES {
 			let source = Source::anonymous(content);
 			let template = Template::parse(source)?;
 
 			assert_eq!(
-				&template.resolve::<Variables, Variables>(None, None)?,
-				should
+				&template.resolve::<Variables, Variables>(Some(&vars), None)?,
+				should,
+				"Format test failed for input `{}`",
+				content.escape_debug()
 			);
 		}